refactor: Rearrange config files for firebase, update naming conventions
